Rename the util._select module to util._selector
This is to match the main contents of the module, the `Selector` class.
