+++ src/frequenz/channels/util/_selector.py
